// build-test/src/types/request.d.ts
export declare namespace RequestParams {
  interface Login {
    name: string;
  }

  interface CreateProductParams {
    name: string;
    remark: string;
  }
  interface UpdateProductParams {
    name: string;
    remark: string;
    id: number;
  }
  interface GetProductList {
    name: string;
    // manager: string;
    pageIndex: number;
    pageSize: number;
  }
  interface UpdateStatus {
    id: number;
    status: number;
  }
  interface IterationParams {
    status: number;
    pageIndex: number;
    pageSize: number;
  }
// ...
  interface UpdateEmployeeList {
    iteration_id: number | null | string;
    staff_no_list: Array<string>;
  }

  interface UpdateAppList {
    iteration_id: number | null;
    list: Array<Record<string, any>>;
  }

  interface UpdateDoc {
    quote_id: number;
    category: number;
    /**
     * 1:链接；2:文件
     */
    type: number;
    url: string;
  }
  interface CreateIteration {
    notify_url: string;
    backend_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    dev_time: string | null;
    frontend_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    id: number | null;
    iteration_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    name: string | null;
    is_schedule: number;
    plan_id: number | null;
    release_time: string | null;
    test_manager: {
      staff_name: string | null;
      staff_no: string | null;
    };
    test_time: string | null;
    union_time: string | null;
  }

  interface GetTaskList {
    pageIndex: number;
    pageSize: number;
    status: number;
    iteration_id: number;
    type: number;
    name: number;
    staff_no: string;
  }

  interface UpdateTaskList {
    complete_percent?: number;
    id: number;
    status: number;
  }

  interface UpdateTaskListLevel {
    ids: number[];
    level: number;
  }

  interface GiveTaskList {
    staff_name: string;
    task_ids: number[];
    staff_no: string;
  }
  interface CreateTaskList {
    begin_time: string;
    end_time: string;
    hours: number;
    iteration_id: number;
    name: string;
    remark: string;
    staff_name: string;
    staff_no: string;
    type: number;
  }
// ...
  interface UserIterationParams {
    month: string;
    pageIndex: number;
    pageSize: number;
  }

  interface GetTestParams {
    name: string;
    status: string;
    product_id: number;
    pageIndex: number;
    pageSize: number;
  }

  interface GetPeopleParams {
    staff_no_list: string[];
    iteration_ids: string[];
    plan_ids: string[];
    product_id: number;
    pageIndex: number;
    pageSize: number;
  }
// ...
  interface GetUseCaseList {
    pageIndex: number;
    pageSize: number;
    status: number[];
    product_id: number;
    type: number[];
    level: number[];
    name: string;
    create_by: string;
  }
  interface GetUseQualityList {
    pageIndex?: number;
    pageSize?: number;
    page_size?: number;
    status?: number;
    product_id: number;
    iteration_id: number;
    start_time: string;
    month: string;
    end_time: string;
  }

  interface RelativeTaskList {
    product_id: number;
    iteration_id: number;
    filter_type: number;
  }
  interface SelectTaskLevel {
    product_id: number;
    iteration_id?: number;
    filter_type?: number;
  }

  interface UpdateIterationDemand {
    iteration_id: number;
    demand_ids: number[];
  }
}